Updated openstack/openstack Project: openstack/heat  919de6acd1e7fd495449cff03d04aaea64b11e92
no need to run integration tests in upgrade

the purpose of upgrade script is to upgrade service,
and heat shouldn't run any jobs inside of that. moreover
there is no benefit in executing these tests after upgrade
because it doesn't test anything related to upgrades:
if heat is correctly configured after upgrade it will be ready
to run tests (but this can be covered by usual integration job).
heat should be testing specific things like
updating stacks after migration to the new openstack code,
deleting stacks and so on.

Co-authored-by: Vitaly Gridnev <vgridnev@mirantis.com>
Change-Id: Ie600dfc09db0ffa91609f92be4b9ef27259b5314