Updated openstack/openstack Project: openstack/tempest  882545e40066dc6f7f5a808ed1b6d4136ae99ea9
Configure stable service clients via the registry

Push the configuration for stable service clients into the client
registry so that they may be automatically configured in clients.py.
This makes the stable service clients available and pre-configured
to both Tempest and all the plugins, removing the need for the
extra configuration step.

Once all tempest service clients will be stable, several pieces of
code will become redundant, as noted in the code itself.

Change-Id: I6a4845edb95031243bca12a8d03c60cf18528212