Updated openstack/openstack Project: openstack-infra/project-config  6e9acf5f38494d25e729254679c9f6bb0aaa3eda
Get IPA buildimage POST job into healthcheck

This patch adds subunit output to the end of the buildimage job so that
heathcheck will have information on success or failure of the job.
This patch implements this using an exit trap, this is a temporary
solution until Zuul can report the success/failure of jobs directly to
healthcheck. This solution will only catch failures that occur during
the actual image build process, any errors that occur inside the scp
publisher's will not be detected by this change.

Change-Id: Id26ed8f40cc659e065c598e2e7af760e8be42419