Updated openstack/openstack Project: openstack/ironic  3d279ed0f318c83b0a47408844632458f0cbb6bf
Fix updating port MAC address for active nodes

Current update fails with HTTP 500 when trying to update a bound Neutron
port. Remove the binding before update and restore it during update.

For nodes in "active" state (or with instance_uuid assigned) only allow
such update in maintenance mode. This is not a breaking change, as previously
it did not work at all.

Change-Id: I356fc0eb3702eb16b7f8e675ea416473c65af7e3
Closes-Bug: #1611744