Updated openstack/openstack Project: openstack/trove  4063bdb3107c8f721c122220afc0a7c15a6949a1
guest_id missing err, switch config_drive default

This changes the default setting for use_nova_server_config_drive
from False to True. By default, nova does not use a config
drive for launched instances. Therefore trove must implicitly
request the config drive be used so that it can inject
guest_info.conf into the guest VM.

This commit also adds an error to guestagent when guest_id
is missing. If the guest_id configuration parameter is not
set at guestagent startup it will throw a RuntimeError. This
case typically occurs because guest_info.conf was
not injected into the guest, or was not included in the set
of configs that guestagent reads at startup.

And finally, this commit adds a section in the 'building guest
images' documentation describing how configuration is injected
into the guest agent.

Change-Id: I084c841472183893a63ca7b70d560f44a2f08901
Partial-Bug: 1609915