Updated openstack/openstack Project: openstack/glance  11cfe49b8f88f68d83028b5920891bb16792da72
Regenerate config files for Newton

During the release many config options are changed or updated. The
sample config files provided in tree (under etc/*) need to be updated
using the autogeneration tool as these files are being used in different
places. Some examples are devstack gates, downstream packagers shipping
with the same (or using defaults from these files), etc.

Hence, before we cut a release we need to refresh the config files
shipped with tree to match the changes done in the source code during
the release cycle.

This commit is intended to refresh the entire set of config files for
Newton. They include:

    * glance api and registry related configs
    * scrubber, cache and manage utilities' configs
    * any other configs coming from downstream like oslo libraries,
      glance_store, etc.
    * belonging to affiliated code base like glare.

Change-Id: Iea2c538b37182191445cc5c1f8b2d9dceed06343
Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>
Co-Authored-By: Hemanth Makkapati <hemanth.makkapati@rackspace.com>