Updated openstack/openstack Project: openstack/tripleo-heat-templates  2c877c9c4b08983ac2d45e1c2c2a8da60f608ad5
Align node_ips hiera keys with the service name.

To enable composable generation of this switch the key names
to align with the service_name of each service.

Note that this should depend on I423b544df174254ac511b906b0c570e701678022
and previously passed CI with that defined, but because we now run
gate validation jobs on puppet-tripleo it's impossible to land, so
this now contains both old and new hiera keys temporarily, which will
be removed when the puppet-tripleo patch lands.

Change-Id: I7febf28bf409e25e8e5961ab551b6d56bb11e0c6
Partially-Implements: blueprint custom-roles