Updated openstack/openstack Project: openstack/trove  2c0f57e5412cba75298602bd164aa946538e3985
Define safe pid-file and socket paths in config

Service manager cannot properly manage the MySQL
process if the pid file / socket are not defined
for the mysql_safe process.

Also rename 'pid_file' to 'pid-file', default for
both 5.6 and 5.7 and the only name accepted in the
'mysql_safe' section.

Update MySQL derivatives: MySQL, Percona and PXC.

Change-Id: I44441144a74e01d3c816ca72b2da4aa351dbd46c
Closes-Bug: 1605385