Updated openstack/openstack Project: openstack/tripleo-heat-templates  8c61510e601473cabd60264eff7fc5b6f9f2ebb8
Add parameter for enabling purging configuration files

This patch adds a parameter, EnableConfigPurge, to the base nova and
neutron service files. The value is passed to their respective puppet
modules to control whether resources not managed by puppet are purged
from the config files. The default value is set to true.

Change-Id: I32732dc2ed7a7b3187e8727592cb4add0c6b9cf6
Closes-Bug:  #1611800