Updated openstack/openstack Project: openstack-infra/tripleo-ci  4c485a0896c89366bc8cc738499ff4980e776410
Reinstall python-requests in bootstrap script

python-requests needs to be reinstalled in bootstrap-overcloud-full.sh
if it was already installed in order to make this script idempotent.
Otherwise, yum sees the package as already installed and does not
reinstall it again after we have rm'd the files (which we do to clean up
the previous pip install).

Change-Id: I6dd6676ae9a53a31afeaec43d242a33e072b4a94