Updated openstack/openstack Project: openstack-infra/shade  d6acf018a37784f2b778d988d388d4937fdc514b
Infer nova-net security groups better

If a cloud does not have neutron, it's entirely possible to infer that
security groups should go through nova. While it's currently possible to
set secgroup_source: nova in clouds.yaml or in a vendor file, it's a
little overkill for folks who otherwise do not need to express that
level of complexity.

Change-Id: I8e4dec2cf993a47c8a0d5cb55bbf80b0fdc4dd46