Updated openstack/openstack Project: openstack-infra/tripleo-ci  dbc8e1fa0f9e8597bd3dc81f33ae387fc5cc536d
Bump baremetal vm memory to 5.5 GB

We're seeing OOM issues on overcloud nodes on a pretty regular
basis now, so it appears we need to increase the amount available.
Hopefully increasing from 5 to 5.5 GB will be sufficient to
eliminate the problem without allowing us to bloat our memory usage
accidentally.

Change-Id: I6781af7629b4c870e735e42e0b7c576b25b13f83
Closes-Bug: 1611842