Updated openstack/openstack Project: openstack-infra/tripleo-ci  433792e2c51489976646ae7578b273266847cf12
Don't try to get stack details if the stack doesn't exist

If a run fails before the overcloud is deployed or we are running
get_host_info.sh after deleting the overcloud then all of the
commands that operate on the stack will fail.  Let's not even run
those if we don't find the stack.

Change-Id: I06ff879424793ea1372889ae2c0e0740c45cb5c8