Updated openstack/openstack Project: openstack-infra/nodepool  346b34b1f84271a3c192ef022f8aa2bf360a1fe5
Make ZK fixture more robust

After accounting for a recent change in ZK log syntax, we were not
correctly detecting when a ZK server hit a port collision on startup.
This corrects that, as well as adding some additional logging,
including outputting the zookeeper server log at shutdown.  We might
not want that in the long run, but it could be handy for now.

Change-Id: I5ad8e375481bda0e2f4da5d93f57406f3c1e9aa9