Updated openstack/openstack Project: openstack/glance_store  28009baa2202387d5ba43f04d86a689068529adb
Remove "Services which consume this" section

In the first few improvements to the help text of configuration
options, the section "Services which consume this" was included
to clearly indicate which services consume a particular
configuration option. However, this is redundant information as
all glance services have their configuration files. And, the very
fact that a configuration option appears in a certain file
indicates that the corresponding service potentially consumes it.

Change-Id: I1dbaf620d5d9e7025a491b28546be66bfa4442c3
Partial-Bug: #1570946