Updated openstack/openstack Project: openstack/nova  51d62ce99faed99476de036a84f999a0c493feca
Maintain backwards compat for listen opts

Technically you can use a hostname, that then gets resolved by DNS to
find the IP address to bind to. Although this is not actually a very
good idea, because it just picks the first IP it gets and tried to bind
to that one, which is not always the best idea.

Related-Bug: #1611940

blueprint centralize-config-options-newton

Change-Id: Icd271b62935969036b6622b8582edec8bd496dd2