Updated openstack/openstack Project: openstack/neutron  1e6b818fdca5cd9638b0dc349288598048f6fd4c
Avoid IPAM driver reusing a session that has been rolled back

With the in-tree pluggable IPAM driver, IPAM rollback tries to use the
DB session after it has been rolled back due to an exception. This
driver doesn't need roll back, so fix this by adding a method to the
driver signalling that rollback shouldn't be attempted.

Change-Id: Ic254789e58a8a51cd1aa943cb71de12410f4c0a7
Closes-Bug: #1603162
Related-Bug: #1516156