Updated openstack/openstack Project: openstack/neutron  ca0d752d81e568337abb6fb0ecce40e51e67cb92
Introduce ovo objects for security groups

Those objects are intentionally not integrated into the database code so
far. This is to quicken access to their definitions to implement
push-notifications for security groups and ports.

Expanded the list of possible values for IpProtocolEnumField to allow
port numbers in addition to symbolic names since security group API
supports both.

This patch also makes DeclarativeObject not override extra filters on
class creation and instead reuse the list of filters defined on the
class as an attribute. This allows to define it per class not per object
(in __init__).

Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db
Partially-Implements: blueprint push-notifications

Change-Id: Idcdc928d37d5d661428525700001154f50c3056e
Co-Authored-By: Sayali Lunkad <sayali.lunkad@suse.com>
Co-Authored-By: Victor Morales <victor.morales@intel.com>