Updated openstack/openstack Project: openstack/heat  a6d8b95db51972c1c1a0c07688e5918fed0c4fc7
Fix path_in_stack for scheduler hints

Nova expects scheduler_hints to be a DictOfListOfStrings in
the object layer(versioned objects), though there is no
limitation for this in the api layer. Thefore sending a
list of tuples as a scheduler_hint results in error.

This also sets stack_scheduler_hints=true for integration
tests.

Change-Id: Id7575f67657fab86acb22d86807127dda45305d5
Closes-Bug: #1608452