Updated openstack/openstack Project: openstack/heat  dfe3f4f7404ff9b31fd11e66c3c5e03bf3cfe90d
Catch NotSupported when cancelling a nested stack

If a nested stack is in a state where we can't cancel it - generally
because it has actually already finished before we attempt to stop it -
then a NotSupported exception is raised. Catch and ignore this exception
so that we can continue (by issuing an update to the previous template,
in the case of a rollback).

Change-Id: Ied4356193e383881d9da33429809339a4f41ac0b
Related-Bug: #1591337