Updated openstack/openstack Project: openstack/cinder  3114010d03008c358377d130e31ebb586f6a8a24
Remove the using of dict.get() in assertIsNone

The assertIsNone(dict.get(key)) will pass in the
following two cases:
A. if there is no key in the dictionary.
B. if the related value of the key is None.

In this patch, we will replace it with
assertNotIn(key, dict) to eliminate these vague
assertion.

Change-Id: I56ff5119f84d17e73011e10660ae62b4ea4dfa3c
Partial-Bug: #1613154