Updated openstack/openstack Project: openstack/ironic  f55de1bf817fd73aa86fae6f0f8848027e2a1f01
DevStack guide: Bump IRONIC_VM_SPECS_RAM to 1280

This patch is bumping the value of the IRONIC_VM_SPECS_RAM configuration
from 1024 to 1280 in the DevStack guide. Currently the CoreOS version of
the deploy ramdisk can not boot with 1024 and attempting to do it will
result in the boot process falling into the emergency shell.

Our documentation should set the minimum value to something that allow
users to boot any of the supported upstream versions of the IPA ramdisk.

Change-Id: I6760a4311cdc9103826c309070c262e685c85e37
Closes-Bug: #1611773