Updated openstack/openstack Project: openstack/cinder  822c804262f8de9656806dc6f2e7b8d3c3a2ee6c
Fix Scheduler manager masking retype fail details

When a retype fails because it requires a migration and we haven't
changed the migration policy -so it has default "never" value- we don't
get a specific log message as we should, instead we get a generic error
because the scheduler manager is hiding the error details creating a
custom error message.

This patch fixes this by logging the real exception raised by
find_retype_host.

Change-Id: Ic086bdcc65780c5b76b671a161fa10a86dfa60c1
Closes-Bug: #1605673