Updated openstack/openstack Project: openstack/python-novaclient  05bfe1fce8e79a299e5410b59d8058775bc7653c
Handle successful response in console functional tests

The console functional tests were written under the assumption
of how the gate jobs work in our CI system, but with a local
devstack you might have a different configuration. For example,
we don't run the nova-novnc service in the gate but it's enabled
by default in devstack, so functional console tests for vnc will
fail in a local default devstack setup.

This change fixes the console functional tests to handle the case
that a valid response is returned and adds a check for how the
response looks so we're at least validating something.

Change-Id: I3505c9955d032301b34764f67466c467a79ed9fb
Closes-Bug: #1613435