Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  d2d7c3f3f982068b0162c7d13b52827a2bdd0619
Use xml_jobs not jobs

Fixes a regression where using --delete-old will delete jobs that are
not updated for example in the case of caching and jenkins-jobs
determines a job does not need to be updated.

jobs returns a list of jobs that were updated via jenkins-jobs while
xml_jobs returns a list of all jobs generated jobs regardless of if they
were updated or not.

Change-Id: I8f9d0b609d1a7cc0a526d479b4f0a2448c25aba3