Updated openstack/openstack Project: openstack-infra/project-config  4ce405eaddd169e88f2020927257170d892f4d92
Ensure we alway build old Ironic ramdisk

Ironic doesn't provide prebuilt images for the old deprecated deploy
ramdisk this must always be built by devstack or manually provided. In
the job we don't manually provide the ramdisk and after the recent
refactor building the ramdisk is explictly disabled by the job this
means devstack fails, because it doesn't have a ramdisk to use. This
patch reenables building the ramdisk for the job that uses the old
ramdisk.

Change-Id: I0f4d90d90dc47f102a52cba242ba9533a11cb1a2