Updated openstack/openstack Project: openstack/trove  27a076f6d6dee97d25f60457f1450e1942c2c4e8
Add missing ICMP option for pxc & mariadb

An option called "icmp" was added to cfg.py to allow for the
ICMP protocol to be added to the security group created by
task manager. The option was added to all datastores except
MariaDB and PXC. This commit corrects that and adds the option
for those 2 datastores.

Change-Id: Icd2845cf689fecb7bda18046061b601b3ee123ed
Closes-bug: 1613768