Updated openstack/openstack Project: openstack-infra/puppet-snmpd  2de225f9331c48acc3600353e3f9202cbf12f599
Remove initscript

This sort of reverts 26d074aba3172d7019a47ada003531b7b7f80584 whose
commit message included the phrase, "This can go away after
everything is upgraded to precise..."

We're well past upgrading to Ubuntu Precise, as we're actively
upgrading the last of our Precise servers to Trusty now. Further,
the initscript we backported from Precise no longer works on Xenial
which we're starting to get into production, and we haven't been
installing this file on our RH-based platforms at all, so it's long
overdue to be cleaned up.

Change-Id: I3f99495142ad1b4352c1566889381fe5c92b2936