Updated openstack/openstack Project: openstack-infra/elastic-recheck  8419a21a4997f5914399e26a6f0d97bba3702e79
Add query for bug 1613749

This commit adds a query for the http timeout for requests to glance
when we ask to use a remote image. There are multiples per run, since
it gets logged twice per test failure and there are 2 or 3 failures
per run when this issue occurs.

Related-Bug: #1613749
Change-Id: Ic3dd31ee9305660e3f788ce3d059f535f2a95c06