Updated openstack/openstack Project: openstack/governance  e53d24cd9b4418a721000710d33c8d1cfc84da4c
describe a process for managing community-wide goals

One of the outcomes of the discussion at the leadership training session
earlier this year was the idea that the TC should set some
community-wide goals for accomplishing specific technical tasks to get
the projects synced up and moving in the same direction.

This patch includes the proposed process for managing those goals, a
template for describing a goal, the placeholder for goals for the Ocata
cycle, and a tool for mechanically generating part of the documents.

I had help with reviews and drafts of some portions of this text from a
bunch of other folks including jroll, dtroyer, johnthetubaguy, spamaps,
mordred, ttx, notmorgan, thingee, and amrith.

Change-Id: I46cc66eafe9dbcfda94fb6a1a8c66c2ca77de45a
Signed-off-by: Doug Hellmann <doug@doughellmann.com>