Updated openstack/openstack Project: openstack/neutron  62243c45b008f9703ebe95178ed9eb0094293158
Fix test_router_add_interface_delete_port_after_failure

In test_router_add_interface_delete_port_after_failure,
when a subnet is created, a neutron DHCP port could be created
if this subnet is DHCP-enabled.

In this case, the test will fail because it only checks if a
router-interface port is deleted and expects no more port in
the system. But it misses the internal DHCP port is still there.

The patch here is to explicitly set enable_dhcp=False when
creating the subnet.

Change-Id: I4acbf2c83f9145f7b99519ec90934fa862de67f7
Closes-Bug: #1612894