Updated openstack/openstack Project: openstack-infra/shade  2d67c978e09e1b00cefec2e58981c12d19142196
Base auto_ip on interface_ip not public_v4

Currently the auto_ip logic checks for a public v4 address to determine
if it should make a FIP. However, we have richer logic that provides
interface_ip that takes in to account IPv6 as well as configured choices
as to what the "default" network should be. That is what should really
be cared about - does the server have an ip that would be considered the
thing to connect to it on.

Change-Id: Ica1feb31cc3477c6336a749c88f4706efd20484f