Updated openstack/openstack Project: openstack-infra/shade  046dc3e164581e4e333eac6b360be8e6cfa30f86
Honor default_interface OCC setting in create_server

If a network has been configured as the "default_interface" in a
clouds.yaml file, then it should be used as the default value to the
nics argument in create_server. If a nics or a network argument are
given to create_server, that should, of course, win.

Change-Id: I278848f958a464daf2a5249084216b013698d62a