Updated openstack/openstack Project: openstack/nova  d23fb5ff9f10559681adc04b5b4116cfb0ede9df
Make simple_cell_setup work when multiple nodes are present

The current code assumes that there is only one node per compute host,
and thus will try to create mappings for the same host if that host
has multiple nodes (i.e. in ironic).

Change to a set so we do it only once per host.

Change-Id: I1619ece851a1a5c090eb70bdca90e396465e328e