Updated openstack/openstack Project: openstack/sahara  62df4993f4466bf806d5f05f42b45acbb66eab1f
Refactor the logic around use of floating ips

The logic associated with the use_floating_ips configuration flag is conflated
and needs to be refactored. Essentially, whether or not sahara uses a floating
ip for instance management should be separate from whether or not an instance
is required to have a floating ip.

Implements bp: refactor-use-floating-ip

Change-Id: I9f666e8f9f476f34a508198b271568db17548565