Updated openstack/openstack Project: openstack/glance_store  14b9e1fdf5efea805ff29b12f328061e4333b87d
Release note for glance_store configuration opts.

Adding a release note stating the work on improving configuration
options of Glance store for this release.
Some of the configuration options in Glance store were set
with a ``min``, preventing them to accept negative integers like
in the previous release. This release note will give operators a
heads up on this.

Change-Id: I8db8764f84c92e987a774ce3650beb53d9635905